assignment-4/requirements.txt
# Assignment 4: Simple RAG Agent - Requirements
# RAG system and embeddings dependencies

openai
python-dotenv
numpy

# Optional: For enhanced functionality
# faiss-cpu  # For better vector search
# chromadb   # Alternative vector database
# streamlit  # For web interface
# PyPDF2     # For PDF document loading

# Development dependencies
# pytest
# black 